Fix choco scripts

+++ chocolatey/tools/chocolateyuninstall.ps1
@@ -3,26 +3,20 @@
 $ErrorActionPreference = 'Stop';
 
 $NoteBarDir = "$env:Programfiles\NoteBar"
+$NoteBarAppData = "$env:APPDATA\NoteBar"
 
-# Get regasm path
-$dotnetPath = [System.Runtime.InteropServices.RuntimeEnvironment]::GetRuntimeDirectory()
-$RegasmPath = "$dotnetPath\RegAsm.exe"
-
-# Unregistre notebar deskband
-& $RegasmPath /u "$NoteBarDir\NoteBar.Toolbar.dll"
-Stop-Process -ProcessName explorer
-Start-Sleep -s 2
-
-# Delete NoteBar AppData folder
-Remove-Item -Path "$env:APPDATA\NoteBar" -Recurse -Force
+# Delete NoteBar AppData directory
+if (Test-Path -Path $NoteBarAppData) {
+    Remove-Item -Path $NoteBarAppData -Recurse -Force
+} 
 
 # Delete EventSource of NoteBar logging
-[System.Diagnostics.EventLog]::DeleteEventSource("NoteBar")
+if ([System.Diagnostics.EventLog]::SourceExists("NoteBar")) {
+    [System.Diagnostics.EventLog]::DeleteEventSource("NoteBar")
+}
 
 # Delete notebar from PATH
 $Path = [Environment]::GetEnvironmentVariable("Path")
 $NewPath = ($Path.Split(";") | Where-Object { $_ -ne $NoteBarDir }) -join ";"
-[Environment]::SetEnvironmentVariable("Path", $NewPath, [System.EnvironmentVariableTarget]::Machine)
-
-# Delete src
-Remove-Item -Path $NoteBarDir -Recurse -Force
+[Environment]::SetEnvironmentVariable("Path", 
+    $NewPath, [System.EnvironmentVariableTarget]::Machine)
